SIMD backend for poseidon252 merkle ops - CPU IMPL
