+++ crates/prover/src/core/backend/cpu/mod.rs
@@ -4,6 +4,8 @@ mod circle;
 mod fri;
 mod grind;
 pub mod lookups;
+#[cfg(not(target_arch = "wasm32"))]
+mod poseidon252;
 pub mod quotients;
 
 use std::fmt::Debug;
