+++ crates/prover/src/core/backend/simd/mod.rs
@@ -2,6 +2,8 @@ use serde::{Deserialize, Serialize};
 
 use super::{Backend, BackendForChannel};
 use crate::core::vcs::blake2_merkle::Blake2sMerkleChannel;
+#[cfg(not(target_arch = "wasm32"))]
+use crate::core::vcs::poseidon252_merkle::Poseidon252MerkleChannel;
 
 pub mod accumulation;
 pub mod bit_reverse;
@@ -15,6 +17,8 @@ pub mod fri;
 mod grind;
 pub mod lookups;
 pub mod m31;
+#[cfg(not(target_arch = "wasm32"))]
+pub mod poseidon252;
 pub mod prefix_sum;
 pub mod qm31;
 pub mod quotients;
@@ -26,3 +30,5 @@ pub struct SimdBackend;
 
 impl Backend for SimdBackend {}
 impl BackendForChannel<Blake2sMerkleChannel> for SimdBackend {}
+#[cfg(not(target_arch = "wasm32"))]
+impl BackendForChannel<Poseidon252MerkleChannel> for SimdBackend {}
